Moves mock setup functions with literals back to tests
I decided it was better to have the literals close to the testing code
